% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mex.R
\name{mex}
\alias{mex}
\title{Model-assisted missing data exploration}
\usage{
mex(...)
}
\arguments{
\item{data}{A dataset that is of class `data.frame`}
}
\description{
Fit models to explore the structure of missing values in a dataset.
}
\details{
This is the main function in the \code{mex} package.  It runs all of the
routines for exploring missing data and returns a \code{mex} object which
contains all of the results.
}

